Rewrite Stats Core Data tests to be cleaner
